+++ src/PanelTraits/Query.php
@@ -51,7 +51,36 @@ trait Query
      */
     public function orderBy($field, $order = 'asc')
     {
+        if ($this->request->has('order')) {
+            return $this->query;
+        }
+
         return $this->query->orderBy($field, $order);
+    }
+
+    /**
+     * Order results of the query in a custom way.
+     *
+     * @param  array $column           Column array with all attributes
+     * @param  string $column_direction ASC or DESC
+     *
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function customOrderBy($column, $columnDirection = 'asc')
+    {
+        if (! isset($column['orderLogic'])) {
+            return $this->query;
+        }
+
+        $this->query->getQuery()->orders = null;
+
+        $orderLogic = $column['orderLogic'];
+
+        if (is_callable($orderLogic)) {
+            return $orderLogic($this->query, $column, $columnDirection);
+        }
+
+        return $this->query;
     }
 
     /**