+++ src/app/Http/Controllers/Operations/Create.php
@@ -14,12 +14,13 @@ trait Create
     public function create()
     {
         $this->crud->hasAccessOrFail('create');
+        $this->crud->setOperation('Create');
 
         // prepare the fields you need to show
         $this->data['crud'] = $this->crud;
         $this->data['saveAction'] = $this->getSaveAction();
         $this->data['fields'] = $this->crud->getCreateFields();
-        $this->data['title'] = trans('backpack::crud.add').' '.$this->crud->entity_name;
+        $this->data['title'] = $this->crud->getTitle() ?? trans('backpack::crud.add').' '.$this->crud->entity_name;
 
         // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
         return view($this->crud->getCreateView(), $this->data);
@@ -35,17 +36,11 @@ trait Create
     public function storeCrud(StoreRequest $request = null)
     {
         $this->crud->hasAccessOrFail('create');
+        $this->crud->setOperation('Create');
 
         // fallback to global request instance
         if (is_null($request)) {
             $request = \Request::instance();
-        }
-
-        // replace empty values with NULL, so that it will work with MySQL strict mode on
-        foreach ($request->input() as $key => $value) {
-            if (empty($value) && $value !== '0') {
-                $request->request->set($key, null);
-            }
         }
 
         // insert item in the db