+++ src/app/Http/Controllers/Operations/Update.php
@@ -16,6 +16,7 @@ trait Update
     public function edit($id)
     {
         $this->crud->hasAccessOrFail('update');
+        $this->crud->setOperation('Update');
 
         // get entry ID from Request (makes sure its the last ID for nested resources)
         $id = $this->crud->getCurrentEntryId() ?? $id;
@@ -25,7 +26,7 @@ trait Update
         $this->data['crud'] = $this->crud;
         $this->data['saveAction'] = $this->getSaveAction();
         $this->data['fields'] = $this->crud->getUpdateFields($id);
-        $this->data['title'] = trans('backpack::crud.edit').' '.$this->crud->entity_name;
+        $this->data['title'] = $this->crud->getTitle() ?? trans('backpack::crud.edit').' '.$this->crud->entity_name;
 
         $this->data['id'] = $id;
 
@@ -43,17 +44,11 @@ trait Update
     public function updateCrud(UpdateRequest $request = null)
     {
         $this->crud->hasAccessOrFail('update');
+        $this->crud->setOperation('Update');
 
         // fallback to global request instance
         if (is_null($request)) {
             $request = \Request::instance();
-        }
-
-        // replace empty values with NULL, so that it will work with MySQL strict mode on
-        foreach ($request->input() as $key => $value) {
-            if (empty($value) && $value !== '0') {
-                $request->request->set($key, null);
-            }
         }
 
         // update the row in the db