+++ src/resources/views/fields/upload.blade.php
@@ -7,7 +7,11 @@
     @if (!empty($field['value']))
     <div class="well well-sm">
         @if (isset($field['disk']))
+        @if (isset($field['temporary']))
+            <a target="_blank" href="{{ (asset(\Storage::disk($field['disk'])->temporaryUrl(array_get($field, 'prefix', '').$field['value'], Carbon\Carbon::now()->addMinutes($field['temporary'])))) }}">
+        @else
             <a target="_blank" href="{{ (asset(\Storage::disk($field['disk'])->url(array_get($field, 'prefix', '').$field['value']))) }}">
+        @endif
         @else
             <a target="_blank" href="{{ (asset(array_get($field, 'prefix', '').$field['value'])) }}">
         @endif