+++ src/resources/views/form_content.blade.php
@@ -3,11 +3,13 @@
 @endif
 
 {{-- See if we're using tabs --}}
-@if ($crud->tabsEnabled())
+@if ($crud->tabsEnabled() && count($crud->getTabs()))
     @include('crud::inc.show_tabbed_fields')
     <input type="hidden" name="current_tab" value="{{ str_slug($crud->getTabs()[0], "") }}" />
 @else
+    <div class="box col-md-12 padding-10 p-t-20">
     @include('crud::inc.show_fields', ['fields' => $fields])
+    </div>
 @endif
 
 {{-- Define blade stacks so css and js can be pushed from the fields to these sections. --}}